Publish to s3 support.
The first time run this command it'll ask you to enter the docs AWS creds. After that you shouldn't have to do it anymore.

Also, this rake task uses s3cmd which you can brew install or follow the link in the error output. That tool seems to beat any ruby library out currently to accomplish this task.